+++ lib/screens/loading.dart
@@ -16,6 +16,7 @@ import 'package:latlong2/latlong.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:country_coder/country_coder.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 class LoadingPage extends ConsumerStatefulWidget {
   @override
@@ -24,6 +25,7 @@ class LoadingPage extends ConsumerStatefulWidget {
 
 class _LoadingPageState extends ConsumerState<LoadingPage> {
   String? message;
+  static const kPrefLastSizeWarning = 'last_size_warning';
 
   Future doInit() async {
     final loc = AppLocalizations.of(context)!;
@@ -67,6 +69,20 @@ class _LoadingPageState extends ConsumerState<LoadingPage> {
     LatLng? location = ref.read(geolocationProvider);
     if (location != null) {
       ref.read(effectiveLocationProvider.notifier).set(location);
+    }
+
+    // Alert if there are too many changes loaded.
+    final needSizeAlert =
+        ref.read(osmDataProvider).length >= kMinElementsForWarning;
+    if (needSizeAlert) {
+      final prefs = await SharedPreferences.getInstance();
+      if (DateTime.now().day != prefs.getInt(kPrefLastSizeWarning)) {
+        await prefs.setInt(kPrefLastSizeWarning, DateTime.now().day);
+        AlertController.show(
+            'Too much data',
+            'Purge downloaded data from Settings to make the editor faster.',
+            TypeAlert.warning);
+      }
     }
 
     // Finally switch to the monitor page.